requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
boto==2.49.0
boto3==1.22.8
botocore==1.25.8
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.0
gunicorn==20.1.0
jmespath==1.0.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.10.0
whitenoise==5.3.0
